+++ src/main/java/lotto/domain/lotto/LottoAutomaticTicket.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.List;
 
@@ -10,6 +10,7 @@ public class LottoAutomaticTicket extends Lotto{
         super(numbers);
         this.numbers = numbers;
     }
+
     public List<Integer> getLotto(){
         return this.numbers;
     }
